% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runIR.R
\name{runIR}
\alias{runIR}
\title{Shiny app with exploratory analysis of the term structure of interest rates}
\usage{
runIR()
}
\value{
Runs the Shiny app retreiving the data, running the analysis and showing the report with interactive visualisations.
}
\description{
The function works with no arguments. The function depends on the dataset \code{\link{term_structure}}
}
\seealso{
\code{\link{term_structure}}
}
