% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{term_structure}
\alias{term_structure}
\title{US dollar LIBOR interbank rates dataset}
\format{tibble}
\usage{
term_structure
}
\description{
A dataset containing  US dollar LIBOR interbank rates, observed at monthly frequency, 
for rates spanning the period 1961-2008
}
\keyword{datasets}
